TCPIPHDR_DELTA: Fix potential negative value
sizeof() returns a size_t so the tcpiphdr / ip+tcphdr difference will be
a size_t and always be >= 0, while this intended to detect the
difference getting < 0.

This is actually a no-op with the current code because it currently has
tcpiphdr bigger than ip+tcphdr.

Spotted by Coverity: CID 212435.
Spotted by Coverity: CID 212440.

Signed-off-by: Samuel Thibault <samuel.thibault@ens-lyon.org>