meson: remove meson-dist script
Unfortunately meson subprojects do not support add_dist_script, so we
cannot generate the .tarball-version file at "meson dist" time.
Include the version in the meson project declaration, and use
build-aux/git-version-gen only to determine SLIRP_VERSION_STRING.

Instead of the dist script, we check that the version in the
project declaration matches the latest tag.  If they do not match
it will be impossible to run "ninja dist" successfully.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>