meson: support compiling as subproject
Skip installation of devel files if compiling as
a subproject, including the library if a static version is
available; the parent project can force usage of the (installed)
shared library using default_library=shared.

An option can also be used to customize the SLIRP_VERSION_STRING
and ascertain if the parent project is using the embedded version
of slirp.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>