Remove alloca() call in get_dns_addr_resolv_conf()
The ALLOCA(3) man-page mentions its "use is discouraged".

For now get_dns_addr_resolv_conf() is called with pointer to
a in_addr/in6_addr structure, and its size. Declare a union
of these structures on the stack, able to hold both of them.
This allows us to remove the alloca() call, keeping the buffer
on the stack.

Add an assertion in the unlikely case another inet address
is handled by this function.

Signed-off-by: Philippe Mathieu-Daudé <philmd@redhat.com>
Message-Id: <20210507133212.1952121-1-philmd@redhat.com>