introduce slirp_timer_new wrapper
In order to support CFI, we would like to avoid passing function pointers out of
libslirp, and instead use opaque integers for the timer callbacks.  This patch
introduces an internal API for creating a timer that is based on such a timer
callback.

For now, it receives the id (SLIRP_TIMER_RA is the only one) and calls
slirp->cb->timer_new.  The prototype of ra_timer_handler is changed slightly
to fit better with the next patch which introduces a new libslirp function,
to be called on timer expiration.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>