Add VMS_END marker
The migration VMSTATE_END_OF_LIST marker relies on a NULL field
to mark th eend of a list;  that's a bit delicate when the marker is
forgotten and the code hits a NULL by luck.

Use an explicit flag to mark the end as well.

See: https://lists.gnu.org/archive/html/qemu-devel/2022-01/msg03025.html

Signed-off-by: Dr. David Alan Gilbert <dgilbert@redhat.com>