slirp: invoke client callback before creating timers
The introduction of .timer_new_opaque adds an interesting conundrum.
The Slirp* needs to be stored in .timer_new_opaque so that it can be
passed back to slirp_handle_timer, but it is not returned by slirp_new
and slirp_init until after the first call to .timer_new_opaque (which
is in ip6_init).  This is a problem for programs that, like QEMU, use
more than one Slirp*.

Fix them by passing the Slirp* to a callback before slirp_new returns,
and initializing the timer afterwards.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>