ncsi: Use response header for payload length
In the Linux NC-SI driver[1], each response's length is validated with a
statically declared payload length, _unless_ it's an OEM command or some
more complicated NC-SI packet that you can't determine the length of just
from the "type" field, in which case it just uses the length provided by the
response's header.

To support OEM response handlers without requiring too many modifications we
can make the default payload length use the value specified in the handler
table, and then allow OEM handlers to override the length by modifying the
"length" in the response header within the handler implementation.

[1] https://github.com/torvalds/linux/blob/ec7f49619d8ee13e108740c82f942cd401b989e9/net/ncsi/ncsi-rsp.c#L1215-L1220

Signed-off-by: Peter Delevoryas <pdel@fb.com>