Use target address from Neighbor Advertisement
The IPv6 support in libslirp fails to work with any NDP proxy.  The code
used to interpret the NA packets uses the wrong address to insert into
it's neighbor table, the address of the source of the packet, aka the
proxy itself.

However the NA packet got the real target address readily available.
Just use it directly instead.

Signed-off-by: Bastian Blank <waldi@debian.org>